// frontend/src/generated/model/dueMemoryPromptResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoryPromptResponse } from "./memoryPromptResponse";

export interface DueMemoryPromptResponse {
  overdue_seconds: number;
  prompt: MemoryPromptResponse;
  reason: string;
}